Intent: Fortify operator registry coherence
Operators involved: [Emision, Recepcion, Coherencia, Disonancia, Acoplamiento, Resonancia, Silencio, Expansion, Contraccion, Autoorganizacion, Mutacion, Transicion, Recursividad]

Affected invariants: [#4, #7]

Key changes:\n- Move structural operator classes into tnfr.operators.definitions with preserved metadata\n- Centralize OPERADORES mapping and validation helpers in dedicated modules\n- Update structural API and operator package exports to reference new layout

Expected risks/dissonances: Minimal; import graph shuffled but validation and execution semantics untouched

Metrics: C(t), Si, nu_f unchanged; validation smoke tests cover regressions

Equivalence map: tnfr.structural continues exporting operators and validate_sequence
