+++ src/tnfr/validation/__init__.py
@@ -1,0 +1,5 @@
+"""Validation helpers for TNFR sequences."""
+
+from .syntax import validate_sequence
+
+__all__ = ("validate_sequence",)
